Note env variables for cli color
Argh mangles the formatting of this.
I would like the formatting to be like,
```
...
Note:
Color can be prohibited with the enviromental variable NO_COLOR.
Color can be forced using the enviromental variable CLICOLOR_FORCE.
```

Right now it just does this
`... Note: Color can be prohibited with the enviromental variable NO_COLOR. Color can be forced using the enviromental variable CLICOLOR_FORCE.`
This seems to be a known issue with the argh library, there
is a pull request for this but I don't know when this will
enter crates.io.

Rather unfortunately at the end of the comments for the issue
that discusses this, the author states they have switched
to clap.
I hope I won't have to do the same.